Title: 약수 출력하기, Time: 3063ms, Memory: 9MB, Status: Time Limit Exceed - CodeTree

+++ 231128/약수 출력하기/print-divisor.java
@@ -6,7 +6,7 @@ public class Main {
         
         for(int i=0;i<=N;i++){
             if((double)N%i==0)
-            System.out.print(i+" ");
+            System.out.print((int)i+" ");
         }
     
     }
